refactor(ai): move the model registry into vailabel-models (ai slice 2)
Second slice of the ai-module migration. The static model catalog
(src/domain/ai/registry.rs) is pure data + serde_json, so it moves wholesale
into vailabel-models::domain::registry: RegistryModel, the REGISTRY catalog,
registry_json(), and find(). The binary registry.rs becomes a re-export shim so
commands.rs (ai_model_registry) and plugin.rs (super::registry::find) compile
unchanged.

plugin.rs stays in the binary for now — its ModelPlugin trait/dispatch is
coupled to InferenceAnnotationDraft, AppError, and the feature-gated SAM engine;
its pure prompt types migrate with the AiService restructure (ai slice 3).
crates/models stays FULLY_PURE; +1 crate test.

Green: cargo test -p vailabel-models (7), arch-tests (4), default cargo check.
